refactor: CommonModal로 모달 창 분리
